+++ src/huff/HuffmanHeader.java
@@ -7,21 +7,21 @@ import java.util.HashMap;
 
 public class HuffmanHeader {
 	
-	public static void write(DataOutputStream dos,HashMap<Byte,Integer> freq) throws IOException{
+	public static void write(DataOutputStream dos,HashMap<Byte,Long> freq) throws IOException{
 		dos.writeInt(freq.size());
 		for(byte currentbyte:freq.keySet()) {
 			dos.writeByte(currentbyte);
-			dos.writeInt(freq.get(currentbyte));
+			dos.writeLong(freq.get(currentbyte));
 		}
 	}
 	
-	public static HashMap<Byte,Integer> read(DataInputStream dis) throws IOException{
-		HashMap<Byte,Integer> freq=new HashMap<>();
+	public static HashMap<Byte,Long> read(DataInputStream dis) throws IOException{
+		HashMap<Byte,Long> freq=new HashMap<>();
 		int unique=dis.readInt();
 		
 		for(int i=0;i<unique;i++) {
 			byte currentbyte=dis.readByte();
-			int freqofbyte=dis.readInt();
+			long freqofbyte=dis.readLong();
 			freq.put(currentbyte,freqofbyte);
 		}
